% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debarcode_funcs.R
\name{barcode.assignment}
\alias{barcode.assignment}
\title{Generate barcode assignments based on the result of \code{FlowSOM::SOM}}
\usage{
barcode.assignment(dat, barcode.dims = NULL, m = 3, delta = 0.2)
}
\arguments{
\item{dat}{A \code{data.table} of transformed .fcs data containing the markers used to barcode individual samples; usually CD45 for 'live-cell' barcoding. A 'barcode_node' column must also be present -- the result of \code{FlowSOM::SOM$mapping}.}

\item{barcode.dims}{Unless defined, \code{barcode.dims} will include any 'CD45' columns -- those used for 'live-cell' barcoding and as input for \code{FlowSOM::SOM}}

\item{m}{Numeric. Default = 3. Argument passed to \code{utils::combn(x,m)}; used to define the barcode 'scheme': x-choose-m}

\item{delta}{Numeric. Default = 0.2. Defines the delta/minimum distance between lowest positive barcode marker and highest negative barcode marker; used when assigning barcodes to 'ambiguous' nodes -- those that do not have exactly 3 positive markers by the 'valley' metric.}
}
\value{
Numeric vector of barcode assignments.
}
\description{
Generate barcode assignments based on the result of \code{FlowSOM::SOM}
}
