% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debarcode_funcs.R
\name{barcode.sample.plots}
\alias{barcode.sample.plots}
\title{Plot bivariate pairs showing sample-specific de-barcoding results}
\usage{
barcode.sample.plots(dat, out.name, barcode.ids, bin.number = 100)
}
\arguments{
\item{dat}{a \code{data.table} with a 'barcode' column}

\item{out.name}{Character string for use in plot title; usually the 'batch' name}

\item{barcode.ids}{Named character vector; names should equal that of all unique barcode values, including zero; elements should be some form of unique identifier (usually 'sample.id')}

\item{bin.number}{Numeric; argument passed to \code{ggplot2::geom_hex(...,bins=bin.number)}}
}
\value{
a list of \code{gridExtra} objects containing arranged \code{ggplot2} objects
}
\description{
Plot bivariate pairs showing sample-specific de-barcoding results
}
