% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debarcode_funcs.R
\name{barcode.yield.plot}
\alias{barcode.yield.plot}
\title{Plot barcode yields}
\usage{
barcode.yield.plot(dat, out.name)
}
\arguments{
\item{dat}{a \code{data.table} with a 'barcode' column}

\item{out.name}{Character string for use in plot title; usually the 'batch' name}
}
\value{
a printable list of \code{ggplot2} objects; barcode-specific yields: absolute number and '\%'
}
\description{
Plot barcode yields
}
