% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debarcode_funcs.R
\name{barcode0.node.plots}
\alias{barcode0.node.plots}
\title{Plot bivariate pairs showing convoluted de-barcoding results}
\usage{
barcode0.node.plots(dat0, out.name, bin.number = 100)
}
\arguments{
\item{dat0}{a \code{data.table} with 'barcode' and 'barcode_node' columns; a subset of data contaning only barcode 'zero' values \code{dat[barcode==0]}}

\item{out.name}{Character string for use in plot title; usually the 'batch' name}

\item{bin.number}{Numeric; argument passed to \code{ggplot2::geom_hex(...,bins=bin.number)}}
}
\value{
a list of \code{gridExtra} objects containing arranged \code{ggplot2} objects
}
\description{
Plot bivariate pairs showing convoluted de-barcoding results
}
