% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asinh_viewer.R
\name{cofactor.viewer}
\alias{cofactor.viewer}
\title{Interactive asinh transformation of cytometry data}
\usage{
cofactor.viewer(dat)
}
\arguments{
\item{dat}{\code{data.table} containing raw, untransformed cytometry data}
}
\value{
\code{shinyapp}; \code{asinh.cofactors} - a named vector of cofactor values for use in transforming data - will be assigned to the environment if the \code{actionButton} is clicked.
}
\description{
Interactive asinh transformation of cytometry data
}
