% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dt.R
\name{density.dt}
\alias{density.dt}
\title{Density distributions of selected columns from a \code{data.table}}
\usage{
\method{density}{dt}(dt, sd.cols = NULL, by.cols = NULL, trim.quantile = NULL)
}
\arguments{
\item{dt}{\code{data.table} containing numeric columns and at least one factored column.}

\item{sd.cols}{Character vector; used to define \code{data.table}'s \code{.SDcols} argument; if left undefined, will calculate density distributions for all numeric columns.}

\item{by.cols}{Character vector; used to define \code{data.table}'s \code{by} argument; if left undefined, will group by all non-numeric columns.}

\item{trim.quantile}{Numeric vector of length 2; used to define \code{stats::quantile}'s \code{probs} argument; trims 'extreme events' before calculating density distributions.}
}
\value{
A \code{data.table} with pre-calculated 'density.x' and 'density.y' columns; used for plotting.
}
\description{
Density distributions are pre-calculated using \code{data.table} methods; usally passed to \code{ggplot2} for plotting.
}
