% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_funcs.R
\name{dt.to.fcs}
\alias{dt.to.fcs}
\title{Convert a \code{data.table} into a new .fcs file}
\usage{
dt.to.fcs(
  dt,
  fcs.files.dt,
  parse.by = "sample.id",
  type.convert = NULL,
  name.fix = NULL,
  cyto.method = "spectral",
  write = T,
  out.dir = "./data_modified"
)
}
\arguments{
\item{dt}{Object returned from \code{fcs.to.dt}.}

\item{fcs.files.dt}{Object returned from \code{get.fcs.file.dt}.}

\item{parse.by}{Character string; a named column in \code{dt} that contains unique identifiers; used to both index the \code{dt} and generate individual, unique .fcs file names.}

\item{type.convert}{Argument as defined in \code{parms.adf.dt}}

\item{name.fix}{Argument as defined in \code{parms.adf.dt}}

\item{cyto.method}{Argument as defined in \code{parms.list.from.adf}}

\item{write}{Logical; default \code{TRUE}. A new .fcs file will be written to \code{out.dir}, using the 'batch' column from \code{dt} as a sub-directory.}

\item{out.dir}{File path; default "./data_modified". As files will be modified with both new data columns (\code{FlowSOM} derived) and keywords/meta-data, they should be written to a "./data_modified" folder so that the source is left intact.}
}
\value{
If \code{write} is set to \code{FALSE}, a \code{flowCore::flowFrame} will be returned; otherwise, a new .fcs file will be written.
}
\description{
After reading in (usually large amounts of) .fcs data and converting to a \code{data.table}, the process can be reversed and a new .fcs file created; new keywords are written to the header section and parameters updated using \code{flowCore} functions.
}
