% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_funcs.R
\name{fcs.from.dt.masscyto}
\alias{fcs.from.dt.masscyto}
\title{Convert a \code{data.table} of mass cytometry .fcs data into a new .fcs file}
\usage{
fcs.from.dt.masscyto(
  dt.data,
  reverse.asinh.cofactor = NULL,
  keywords.to.add = NULL,
  fil = NULL,
  out.path = NULL
)
}
\arguments{
\item{dt.data}{A \code{data.table} of .fcs data, as returned from \code{data.table::as.data.table(flowCore::read.FCS(...))}.}

\item{reverse.asinh.cofactor}{If defined, will reverse the \code{asinh} transformation - using the same cofactor value - typically applied to mass cytometry .fcs data}

\item{keywords.to.add}{A list of keywords to append to the header section of the .fcs file; if \code{NULL}, returns a 'basic' \code{flowCore::flowFrame}.}

\item{fil}{If defined, will update the '$FIL' keyword; when writing the new .fcs file, '$FIL' is used to define the \code{flowCore::write.FCS(filename)} argument.}

\item{out.path}{A \code{file.path}; will write the new .fcs to the defined directory; if \code{NULL}, returns an 'updated' \code{flowCore::flowFrame}.}
}
\value{
Depending on defined arguments, returns a \code{flowCore::flowFrame} with/without updated keywords or writes the .FCS to a defined directory.
}
\description{
After reading in (usually large amounts of) .fcs data and converting to a \code{data.table}, the process can be reversed and a new .fcs file created; new keywords are written to the header section and parameters updated using \code{flowCore} functions.
}
