% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_funcs.R
\name{fcs.to.dt}
\alias{fcs.to.dt}
\title{reads a .fcs file and converts it to \code{data.table}}
\usage{
fcs.to.dt(fcs.dt, channel_alias = NULL)
}
\arguments{
\item{fcs.dt}{a \code{data.table} as returned from \code{get.fcs.file.dt}}

\item{channel_alias}{as returned from \code{get.fcs.channel.alias}}
}
\value{
a \code{data.table} of raw, un-transformed numeric expression values with character/factor identifier columns
}
\description{
reads a .fcs file and converts it to \code{data.table}
}
