% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsom_funcs.R
\name{fsom.codes.dt}
\alias{fsom.codes.dt}
\title{Update a \code{FlowSOM} object with a \code{data.table} of clustered codes/SOMs.}
\usage{
fsom.codes.dt(
  fsom,
  append.name = NULL,
  k = NULL,
  umap.codes = F,
  seed.val = 1337
)
}
\arguments{
\item{fsom}{A \code{FlowSOM} object as returned from \code{som}.}

\item{append.name}{Character string; will be used to append the otherwise generically named 'node' and 'cluster' columns.}

\item{k}{Numeric; if defined, will generate \code{k} clusters using \code{FlowSOM::metaClustering_consensus()}.}

\item{umap.codes}{Logical: default \code{FALSE}; if \code{TRUE}, the codes/SOMs will be embedded using \code{uwot::umap()}.}

\item{seed.val}{An elite random seed value used to control otherwise random starts.}
}
\value{
A list appended \code{FlowSOM}object updated with a \verb{$code.dt} containing clusters.
}
\description{
Update a \code{FlowSOM} object with a \code{data.table} of clustered codes/SOMs.
}
