% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsom_funcs.R
\name{fsom.codes.to.clusters}
\alias{fsom.codes.to.clusters}
\title{Cluster FlowSOM codes; calculate per-cluster median expression values}
\usage{
fsom.codes.to.clusters(fsom, k = 20, seed = 1337, cluster.col = NULL)
}
\arguments{
\item{fsom}{FlowSOM object, as returned from \code{FlowSOM::SOM}}

\item{k}{Number of clusters; argument passed to \code{FlowSOM::metaClustering_consensus}}

\item{seed}{Numerical value; sets random seed; argument passed to \code{FlowSOM::metaClustering_consensus}}

\item{cluster.col}{Optional character string; if defined, will rename the otherwise generically named 'cluster' column in the median results}
}
\value{
FlowSOM object appended with list-structured results
}
\description{
Cluster FlowSOM codes; calculate per-cluster median expression values
}
\details{
\code{fsom.codes.to.clusters} uses \code{fsom$codes} to first generate metaclusters, then calculates per-cluster median expression values for use primarily in heatmaps/visualizations. If factored clusters already exist (\code{fsom$cluster$fac}), medians will be updated -- useful following \code{fsom.merge.codes}.
}
