% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_map_clusters.R
\name{generate_fsom}
\alias{generate_fsom}
\title{Build a self-organizing map (wrapper)}
\usage{
generate_fsom(
  dat,
  dims,
  scale.dims = T,
  subsample.val = NULL,
  seed.val = 1337,
  ...
)
}
\arguments{
\item{dat}{Data.table returned from \code{readFCS_dt()}}

\item{dims}{Character vector of dimensions (data.table names/columns) to use when building the SOMs}

\item{scale.dims}{Logical. By default, \code{TRUE}; will scale input data before building the SOMs}

\item{subsample.val}{Optional numeric value defining the number of rows to sample from the input data; sample-specific. The default is \code{NULL}}

\item{seed.val}{Optional random seed value}

\item{...}{Additional arguments for \code{FlowSOM::SOM(...)}}
}
\value{
As from \code{FlowSOM::SOM()}: "A list containing all parameter settings and results"
}
\description{
A wrapper for \code{FlowSOM::SOM()} with a few \code{data.table} sensibilities for scaling/sampling
}
