% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debarcode_funcs.R
\name{get.barcode.key}
\alias{get.barcode.key}
\title{Generate a barcode key}
\usage{
get.barcode.key(barcode.dims, m = 3)
}
\arguments{
\item{barcode.dims}{Character vector}

\item{m}{argument passed to \code{utils::combn(x,m)}}
}
\value{
a barcode key (matrix)
}
\description{
Generate a barcode key
}
