% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_funcs.R
\name{get.fcs.file.dt}
\alias{get.fcs.file.dt}
\title{Get a \link[data.table]{data.table} of .fcs file paths and related metadata.}
\usage{
get.fcs.file.dt(fcs.file.paths, factor.cols = NULL)
}
\arguments{
\item{fcs.file.paths}{Character string; path(s) usually returned from \code{list.files(...,full.names=T,pattern=".fcs")}.}

\item{factor.cols}{Character string; column names to be converted to factor.}
}
\value{
a \link[data.table]{data.table} of full length .fcs file paths and related metadata (parsed from the TEXT header).
}
\description{
Essentially a wrapper around \link{get.fcs.keywords.metadata} with some convenience formatting and class conversions. This structured \link[data.table]{data.table} can be considered the starting point of an analysis workflow. Naming conventions can be enforced by modifying any sample/subject-specific columns; it's not uncommon for typos to occur during the course of a large study and they can be corrected after the keyword metadata has been parsed.
}
\examples{
extdata<-system.file("extdata",package="SOMnambulate")
fcs.files <- list.files(extdata,full.names=TRUE,pattern=".fcs")

fcs.files.dt<-get.fcs.file.dt(fcs.files)

#enforce/establish naming convention for the construction of a sample.id and associated factors
new.cols<-c('study.name','batch.seq','batch.date')
fcs.files.dt[,(new.cols) :=
data.table::tstrsplit(basename(dirname(FILENAME)),"_",type.convert=list(as.factor=1:3))]
fcs.files.dt[,stim.condition := factor(stringr::str_extract(FIL,"SEB|UNSTIM"))]
fcs.files.dt[,aliquot.seq := factor(seq(.N)),by=.(batch.date,stim.condition)]
fcs.files.dt[,batch := factor(paste(study.name,batch.seq,batch.date,sep="_"))]
fcs.files.dt[,sample.id := paste(batch,stim.condition,aliquot.seq,sep="_")]
fcs.files.dt[]

}
