% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_funcs.R
\name{get.fcs.keywords.metadata}
\alias{get.fcs.keywords.metadata}
\title{Get keyword metadata (!'$P') from .fcs file header(s)}
\usage{
get.fcs.keywords.metadata(fcs.file.paths, return.dt = F)
}
\arguments{
\item{fcs.file.paths}{Character string; path(s) usually returned from \code{list.files(...,full.names=T,pattern=".fcs")}.}

\item{return.dt}{Logical. By default, \code{FALSE}; if \code{TRUE}, will return a data.table of keyword/values per .fcs file}
}
\value{
a list of non-parameter ('$P') keywords per .fcs file; if \code{return.dt=T}, a data.table of keywords/values
}
\description{
Get keyword metadata (!'$P') from .fcs file header(s)
}
