% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_funcs.R
\name{get.fcs.parameters}
\alias{get.fcs.parameters}
\title{Get parameters ('$P') from .fcs file headers}
\usage{
get.fcs.parameters(fcs.file.paths, return.dt = F)
}
\arguments{
\item{fcs.file.paths}{Character string; path(s) usually returned from \code{list.files(...,full.names=T,pattern=".fcs")}.}

\item{return.dt}{Logical. By default, \code{FALSE}; if \code{TRUE}, will return a data.table of parameters per .fcs file}
}
\value{
a list of parameters per .fcs file; if \code{return.dt=T}, a data.table of parameters
}
\description{
Get parameters ('$P') from .fcs file headers
}
