% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFCS_dt.R
\name{get.parameters}
\alias{get.parameters}
\title{Get parameters ('N'ames and 'S'tains; P$N, P$S) from a .fcs file header}
\usage{
get.parameters(fcs.file.path)
}
\arguments{
\item{fcs.file.path}{Character string; path(s) returned from \code{list.files(...,full.names=T,pattern=".fcs")}.}
}
\value{
a list; \code{[['N']]} contains channel names ($PN); \code{[['S']]} contains channel stains ($PS)
}
\description{
Get parameters ('N'ames and 'S'tains; P$N, P$S) from a .fcs file header
}
