% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debarcode_funcs.R
\name{get.peaks.and.valley}
\alias{get.peaks.and.valley}
\title{Find peaks and valley(s) from a density distribution}
\usage{
get.peaks.and.valley(x, quantile.trim = T, sd.adjust = NULL)
}
\arguments{
\item{x}{Numeric vector}

\item{quantile.trim}{Logical. By default, \code{TRUE}; trims 'extreme' values.}

\item{sd.adjust}{Numeric; if defined, will adjust the return value by multiples of density value standard deviations.}
}
\value{
a named list: \code{$peaks} (peak values) and \code{$valley} (valley values)
}
\description{
Find peaks and valley(s) from a density distribution
}
