% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debarcode_funcs.R
\name{get.valley}
\alias{get.valley}
\title{Find a density distribution 'valley'}
\usage{
get.valley(x, quantile.trim = T, sd.adjust = NULL)
}
\arguments{
\item{x}{Numeric vector}

\item{quantile.trim}{Logical. By default, \code{TRUE}; trims 'extreme' values.}

\item{sd.adjust}{Numeric; if defined, will adjust the return value by multiples of density value standard deviations.}
}
\value{
Numeric representing the 'valley' value in a density distribution.
}
\description{
Using the second derivative of the differences in a density distribution's y values, both peaks and valleys can be detected; if a single valley is detected, its x value is returned. A standard deviation adjustment is included to 'adjust/shift' the valley -- used during 'debarcoding' to fine-tune assignment results.
}
