% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.bivariate.shiny.R
\name{gg.bivariate.shiny}
\alias{gg.bivariate.shiny}
\title{Interactive \code{shiny::shinyApp}; clickable selection of \code{ggplot2::ggplot} parameters}
\usage{
gg.bivariate.shiny(dt)
}
\arguments{
\item{dt}{a \code{data.table} as returned from \code{fcs.to.dt}}
}
\value{
Runs \code{shiny::shinyApp}
}
\description{
Interactive \code{shiny::shinyApp}; clickable selection of \code{ggplot2::ggplot} parameters
}
