% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{gg.func.bivariate.cluster.silhouette}
\alias{gg.func.bivariate.cluster.silhouette}
\title{A cluster-faceted bivariate ggplot with a silhouette sampled from all data points}
\usage{
gg.func.bivariate.cluster.silhouette(
  dt,
  ...,
  bins = 200,
  fill.limits = c(0, bins/2),
  silhouette.val = 50000
)
}
\arguments{
\item{dt}{a \code{data.table} containing a 'cluster' column}

\item{...}{\code{ggplot2::aes} arguments: x and y (unquoted)}

\item{bins}{argument as defined by \code{ggplot2::geom_hex}}

\item{fill.limits}{Numeric vector of length two used to define the \code{limits} argument of \code{ggplot2::scale_fill_gradient(...)}}

\item{silhouette.val}{Numeric; the number of data points to sample from \code{dt} to create the silhouette}
}
\value{
a \code{ggplot2} object
}
\description{
A cluster-faceted bivariate ggplot with a silhouette sampled from all data points
}
