% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{gg.func.density.overlay}
\alias{gg.func.density.overlay}
\title{Plot overlaid (by factor) density distributions of selected columns from a \code{data.table}}
\usage{
gg.func.density.overlay(dt, sd.cols, by.cols)
}
\arguments{
\item{dt}{\code{data.table} containing numeric columns and at least one factored column.}

\item{sd.cols}{Character vector; used to define \code{data.table}'s \code{.SDcols} argument.}

\item{by.cols}{Character vector; used to define \code{data.table}'s \code{by} argument; usually just a single factored column for use with \code{ggplot2::facet_wrap()}.}
}
\value{
A \code{ggplot2} object; if no LHS assignment, directly prints.
}
\description{
Density distributions are pre-calculated using \code{data.table} methods then passed to \code{ggplot2} for plotting.
}
