% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debarcode_funcs.R
\name{key.barcode}
\alias{key.barcode}
\title{Generate a barcode key}
\usage{
key.barcode(barcode.dims, m = 3)
}
\arguments{
\item{barcode.dims}{Character vector.}

\item{m}{Argument passed to \link[utils]{combn}.}
}
\value{
A barcode key (\link[data.table]{data.table}).
}
\description{
Based on a 'x-choose-m' scheme, a full length barcode key -- all possible combinations -- is generated. For mass cytometry live-cell barcoding, the scheme is usually '6-choose-3' (20 unique combinations) or '7-choose-3' (35 unique combinations).
}
\examples{
dt<-SOMnambulate:::prepared.examples("dt")

barcode.dims<-grep("CD45_",names(dt),value=TRUE)

SOMnambulate:::key.barcode(barcode.dims)[]
}
