% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debarcode_funcs.R
\name{key.codes}
\alias{key.codes}
\title{Generate a 'keyed' \verb{[['codes']]}}
\usage{
key.codes(fsom, m = 3, quantile.probs = NULL)
}
\arguments{
\item{fsom}{Object; the result of \link{som.codes.dt}.  The \verb{[['codes']]} result will be used to generate a 'code-specific' barcode key.}

\item{m}{Argument passed to \link{key.barcode}.}

\item{quantile.probs}{Argument passed to \link{get.local.maxima.minima}; Numeric vector of length 2; used to define lower/upper quantile probabilities; trims 'extreme events' before calculating the density distribution.}
}
\value{
A merged \link[data.table]{data.table}; the 'keyed' \verb{[['codes']]} are merged with the result of \link{key.barcode}. Of primary importance is the 'barcode' column which is used to assign barcode values to cellular events (primary data).
}
\description{
Based on the 'x-choose-m' scheme used to barcode individual samples, this function will generate a unique barcode value for each SOM; the barcode value can then be assigned to the primary input data allowing for identification of sample-specific events/cells.
}
\examples{

dt<-SOMnambulate:::prepared.examples("dt")

barcode.dims<-grep("CD45_",names(dt),value=TRUE)
for(j in barcode.dims){data.table::set(dt,i=NULL,j=j,value=asinh(dt[[j]]/10))}
fsom<-som(dt[,barcode.dims,with=FALSE],map=FALSE)
fsom<-som.codes.dt(fsom,append.name = 'barcode')

barcodes<-key.codes(fsom)
barcodes[]

SOMnambulate::map.som.data(fsom,dt)
dt[,barcode := barcodes$barcode[node_barcode]]
dt[,.(node_barcode,barcode)]
dt[,sort(unique(barcode))]
dt[,.N,keyby=.(barcode,stim.condition)]

}
