% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_funcs.R
\name{parms.list.from.adf}
\alias{parms.list.from.adf}
\title{Generate a list of .fcs meta-data}
\usage{
parms.list.from.adf(parms.adf, cyto.method = "spectral")
}
\arguments{
\item{parms.adf}{\code{data.table} returned from \code{parms.adf.dt}}

\item{cyto.method}{A single argument for now; 'spectral'; depending on cytometry platform -- 'spectral','conventional','mass' -- list types might change.}
}
\value{
a list of .fcs meta-data; used to define \code{flowCore::flowFrame(...,description = ...)}.
}
\description{
Generate a list of .fcs meta-data
}
