% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plots.R
\name{plot.pairs}
\alias{plot.pairs}
\title{Generate bivariate plot pairs}
\usage{
\method{plot}{pairs}(plot.dims)
}
\arguments{
\item{plot.dims}{Character vector; a string of dimensions/markers from which bivarate pairs will be generated}
}
\value{
a list of plot pairs; useful for \code{ggplot2} loops
}
\description{
Generate bivariate plot pairs
}
