% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{plotSOMbarcodes_rug}
\alias{plotSOMbarcodes_rug}
\title{Density distribution of FlowSOM nodes per unique barcode channel with barcode-specific nodes as rug-ticks}
\usage{
plotSOMbarcodes_rug(fsom, mdat = NULL)
}
\arguments{
\item{fsom}{An object as returned from \link{som.codes.dt}.}

\item{mdat}{\strong{Optional} data.table of metadata used to associate unique barcodes with unique samples. \code{mdat} must have a 'barcode' column; all additional columns will be used to construct a plot subtitle.}
}
\value{
a \link[ggplot2]{ggplot} object
}
\description{
Density distribution of FlowSOM nodes per unique barcode channel with barcode-specific nodes as rug-ticks
}
\examples{
dt<-SOMnambulate:::prepared.examples('dt')
dt<-dt[stim.condition=="SEB"]
barcode.dims<-grep("CD45_",names(dt),value = TRUE)
for(j in barcode.dims){
data.table::set(dt,j=j,value=asinh(dt[[j]]/10))
}
fsom<-som(dt[,barcode.dims,with=FALSE],.scale=FALSE,xdim=9,ydim=9,map=FALSE)
fsom<-som.codes.dt(fsom,append.name = 'barcode')
SOMnambulate::map.som.data(fsom,dt)
fsom$codes.dt[,barcode := key.codes(fsom)$barcode]
dt[,barcode := fsom$codes.dt$barcode[node_barcode]]

extdata<-system.file("extdata",package="SOMnambulate")
mdat <- data.table::fread(
  list.files(extdata,full.names = TRUE, pattern="batch_sheet.csv"),
  keepLeadingZeros = TRUE
)
mdat<-mdat[stim.condition=="SEB",.(barcode,sample.id,stim.condition,batch)]
plotSOMbarcodes_rug(fsom,mdat)[1:5]

}
