% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{plotSOMbarcodes_zero}
\alias{plotSOMbarcodes_zero}
\title{Bivariate plots of CD45 expression of node-specific barcode 'zero' events}
\usage{
plotSOMbarcodes_zero(
  dt,
  fsom,
  node.threshold = 10000,
  bins.N = 100,
  mdat.cols = NULL
)
}
\arguments{
\item{dt}{An object as returned from \link{fcs.to.dt} and mapped using \link{map.som.data}.}

\item{fsom}{An object as returned from \link{som.codes.dt}; additionally, requires a 'barcode' column.}

\item{node.threshold}{Numeric; to avoid overplotting, if the number of node-specific events exceed the defined threshold value, sampling will take place.}

\item{bins.N}{Argument as defined by \code{ggplot2::geom_hex}.}

\item{mdat.cols}{Character vector; columns in \code{dt} that will be used to construct a plot subtitle.}
}
\value{
a \code{ggplot2} object
}
\description{
Bivariate plots of CD45 expression of node-specific barcode 'zero' events
}
\examples{
dt<-SOMnambulate:::prepared.examples('dt')
dt<-dt[stim.condition=="SEB"]
barcode.dims<-grep("CD45_",names(dt),value = TRUE)
for(j in barcode.dims){
data.table::set(dt,j=j,value=asinh(dt[[j]]/10))
}
fsom<-som(dt[,barcode.dims,with=FALSE],.scale=FALSE,xdim=9,ydim=9,map=FALSE)
fsom<-som.codes.dt(fsom,append.name = 'barcode')
SOMnambulate::map.som.data(fsom,dt)
fsom$codes.dt[,barcode := key.codes(fsom)$barcode]
dt[,barcode := fsom$codes.dt$barcode[node_barcode]]

plotSOMbarcodes_zero(dt,fsom,bins.N=50,mdat.cols=c('batch','stim.condition'))[1:5]
}
