% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{plotSOMclusters}
\alias{plotSOMclusters}
\title{A cluster-faceted bivariate ggplot with a silhouette sampled from all data points}
\usage{
plotSOMclusters(
  dt,
  ...,
  bins = 200,
  show.clusters = NULL,
  silhouette.n = 20000
)
}
\arguments{
\item{dt}{An object as returned from \link{fcs.to.dt} and mapped using \link{map.som.data}.}

\item{...}{\link[ggplot2]{aes} arguments; essentially x and y (unquoted)}

\item{bins}{\link[ggplot2]{geom_hex} argument; numeric vector giving number of bins in both vertical and horizontal directions.}

\item{show.clusters}{Numeric vector; if defined, the faceted plot will show only those clusters.}

\item{silhouette.n}{Numeric (length 1); the number of events to sample from \code{dt} to create the silhouette. Set to 2E4 by default.}
}
\value{
a \link[ggplot2]{ggplot} object; faceted by clusters
}
\description{
Bivariate plot pairs are used to explore FlowSOM clustering results by overlaying cluster-specific data onto a silhouette sampled from all available data; this allows for visualizing how cluster-specific expression relates to the data as a whole.
}
\examples{
dt<-SOMnambulate:::prepared.examples("dt")
fsom<-SOMnambulate:::prepared.examples("fsom")
fsom<-som.codes.dt(fsom,append.name = 'pbmc',k=20)
for(j in colnames(fsom$codes)){data.table::set(dt,i=NULL,j=j,value=asinh(dt[[j]]/10))}
map.som.data(fsom,dt)
plotSOMclusters(dt,CD4_145Nd,CD8a_162Dy,bins=50)
plotSOMclusters(dt,CD4_145Nd,CD8a_162Dy,bins=50,show.clusters=c(4,14))
plotSOMclusters(dt,CD4_145Nd,CD8a_162Dy,bins=50,show.clusters=c(4,14)) +
ggplot2::theme_classic() + ggplot2::guides(fill='none')

}
