% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{plotSOMdensities}
\alias{plotSOMdensities}
\title{Plot density distributions}
\usage{
plotSOMdensities(dt, trim.quantile = c(0.001, 0.999), return.densities = FALSE)
}
\arguments{
\item{dt}{An object as returned from \link{fcs.to.dt}; can be further subset using \link[data.table]{data.table} syntax (see examples).}

\item{trim.quantile}{Numeric of length 2 used to define the \strong{probs} argument of \link[stats]{quantile}; if defined, will 'trim' outlying data before calculating a density distribution.}

\item{return.densities}{Logical; if \code{TRUE}, a \code{data.table} of calculated density distributions (long format) will be returned.}
}
\value{
a (faceted) \link[ggplot2]{ggplot} object
}
\description{
Plot density distributions
}
\examples{
dt<-SOMnambulate:::prepared.examples("dt")
metals <- grep("_[0-9]{3}[A-Z]{1}[a-z]{1}",names(dt),value=TRUE)
for(j in metals){data.table::set(dt,i=NULL,j=j,value=asinh(dt[[j]]/10))}

cols<-c("CD3_170Er","CD4_145Nd","CD8a_162Dy","CD14_146Nd","CD19_142Nd")

plotSOMdensities(dt[stim.condition=="SEB",..cols])
plotSOMdensities(dt[stim.condition=="SEB",.SD,.SDcols=cols,keyby = sample.id])

}
