% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{plotSOMprimary}
\alias{plotSOMprimary}
\title{A bivariate ggplot of hex-binned data points}
\usage{
plotSOMprimary(dt, ..., bins = 200, limits = NULL)
}
\arguments{
\item{dt}{An object as returned from \link{fcs.to.dt}}

\item{...}{\link[ggplot2]{aes} arguments (unquoted); essentially x and y. If a \code{z} aesthetic is provided, the plot will switch to \link[ggplot2]{stat_summary_hex}, using this defined third variable as a 'color-by'.}

\item{bins}{\link[ggplot2]{geom_hex} argument; numeric vector giving number of bins in both vertical and horizontal directions.}

\item{limits}{\link[ggplot2]{continuous_scale} argument.}
}
\value{
a \link[ggplot2]{ggplot} object
}
\description{
A bivariate ggplot of hex-binned data points
}
\examples{
dt<-SOMnambulate:::prepared.examples("dt")
metals <- grep("_[0-9]{3}[A-Z]{1}[a-z]{1}",names(dt),value=TRUE)
for(j in metals){data.table::set(dt,i=NULL,j=j,value=asinh(dt[[j]]/10))}

plotSOMprimary(dt,CD4_145Nd,CD8a_162Dy,bins=100)

plotSOMprimary(dt,CD4_145Nd,CD8a_162Dy,z=CD3_170Er,bins=100)
plotSOMprimary(dt,CD4_145Nd,CD8a_162Dy,z=CD3_170Er,bins=100,
  limits=c(0,dt[,quantile(CD3_170Er,.9)])) +
  ggplot2::theme_dark()
}
