% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{plotSOMprimary_shiny}
\alias{plotSOMprimary_shiny}
\title{A bivariate ggplot of hex-binned data points -- interactive (shiny)}
\usage{
plotSOMprimary_shiny(dt)
}
\arguments{
\item{dt}{An object as returned from \link{fcs.to.dt}}
}
\value{
\link[shiny]{shinyApp}
}
\description{
A bivariate ggplot of hex-binned data points -- interactive (shiny)
}
\examples{
if (interactive()) {
dt<-SOMnambulate:::prepared.examples("dt")
metals <- grep("_[0-9]{3}[A-Z]{1}[a-z]{1}",names(dt),value=TRUE)
for(j in metals){data.table::set(dt,i=NULL,j=j,value=asinh(dt[[j]]/10))}

plotSOMprimary_shiny(dt)
}
}
