% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{plotSOMshiny}
\alias{plotSOMshiny}
\title{Interactive exploration of high-dimensional clustering (FlowSOM) results}
\usage{
plotSOMshiny(dt, fsom)
}
\arguments{
\item{dt}{An object as returned from \link{fcs.to.dt} and mapped using \link{map.som.data}.}

\item{fsom}{An object as returned from \link{som.codes.dt}; fsom$codes.dt is required.}
}
\value{
\link[shiny]{shinyApp}
}
\description{
Interactive exploration of high-dimensional clustering (FlowSOM) results
}
\examples{
if (interactive()) {
dt<-SOMnambulate:::prepared.examples("dt")
fsom<-SOMnambulate:::prepared.examples("fsom")
fsom<-som.codes.dt(fsom,append.name = 'pbmc',k=20,umap.codes=TRUE)
for(j in colnames(fsom$codes)){data.table::set(dt,i=NULL,j=j,value=asinh(dt[[j]]/10))}
map.som.data(fsom,dt)

plotSOMshiny(dt,fsom)
}
}
