% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{plotSOMtime}
\alias{plotSOMtime}
\title{Parameters by Time}
\usage{
plotSOMtime(
  dt,
  trim.quantile = c(0.001, 0.999),
  bins = 200,
  plots.per.page = 5,
  title = NULL,
  timestep = 1e-04
)
}
\arguments{
\item{dt}{An object as returned from \link{fcs.to.dt}; can be further subset using \link[data.table]{data.table} syntax (see examples).}

\item{trim.quantile}{Numeric of length 2 used to define the \strong{probs} argument of \link[stats]{quantile}; if defined, will 'trim' axis (y) limits.}

\item{bins}{\link[ggplot2]{geom_hex} argument; numeric vector giving number of bins in both vertical and horizontal directions.}

\item{plots.per.page}{Numeric; the number of individual plots per page.}

\item{title}{Character string; a plot title (added to each individual page).}

\item{timestep}{Numeric; usually obtained from the '$TIMESTEP' keyword of a .fcs file. The 'Time' parameter is multiplied by the \code{timestep} to convert to seconds with a final conversion to minutes.}
}
\value{
a list (pages) of 'patchworked' (\link[patchwork]{wrap_plots}) \link[ggplot2]{ggplot} objects
}
\description{
Parameters by Time
}
\examples{
dt<-SOMnambulate:::prepared.examples("dt")
metals <- grep("_[0-9]{3}[A-Z]{1}[a-z]{1}",names(dt),value=TRUE)
for(j in metals){data.table::set(dt,i=NULL,j=j,value=asinh(dt[[j]]/10))}

cols<-c("Time","CD3_170Er","CD4_145Nd","CD8a_162Dy","CD14_146Nd","CD19_142Nd")

plotSOMtime(dt[stim.condition=="SEB",..cols],bins=50)[]#'
plotSOMtime(dt[stim.condition=="SEB",c(..cols,'aliquot.seq')],bins=50)[]

}
