% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{plotSOMumap}
\alias{plotSOMumap}
\title{A UMAP of FlowSOM nodes colored by cluster value or expression value}
\usage{
plotSOMumap(
  fsom,
  dt,
  .by = NULL,
  cluster.highlight = NULL,
  expression.facets = FALSE
)
}
\arguments{
\item{fsom}{An object as returned from \link{som.codes.dt}; \code{fsom$codes.dt} is required.}

\item{dt}{An object as returned from \link{map.som.data}}

\item{.by}{Character string; a column in \code{dt} -- usually a factor. \code{.by} is used in \link{dt.N.counts} to generate node counts and in \link[ggplot2]{facet_wrap} to generate facets.}

\item{cluster.highlight}{Numeric; if defined, will highlight specific clusters.}

\item{expression.facets}{Logical; if \code{TRUE}, the UMAP will be faceted by variable and colored by (scaled) expression value.}
}
\value{
a \link[ggplot2]{ggplot} object
}
\description{
A UMAP of FlowSOM nodes colored by cluster value or expression value
}
\examples{
dt<-SOMnambulate:::prepared.examples("dt")
fsom<-SOMnambulate:::prepared.examples("fsom")
fsom<-som.codes.dt(fsom,append.name = 'pbmc',k=20, umap = TRUE)
for(j in colnames(fsom$codes)){data.table::set(dt,i=NULL,j=j,value=asinh(dt[[j]]/10))}
map.som.data(fsom,dt)

plotSOMumap(fsom,dt)

#faceted by 'stim.condition'
plotSOMumap(fsom,dt,.by='stim.condition')

#faceted by 'sample.id'
#for datasets with many samples, this is not feasible
plotSOMumap(fsom,dt,.by='sample.id')

#data.table's 'i' variable to select a single sample
plotSOMumap(fsom,dt[sample.id \%in\% unique(sample.id)[1]],.by='sample.id')

p<-plotSOMumap(fsom,dt[sample.id \%in\% unique(sample.id)[1]],.by='sample.id',cluster.highlight=4)
p
#modify the ggplot object
p + ggplot2::scale_size_area(max_size = 15,guide="none")

#colored by marker expression
plotSOMumap(fsom,dt,expression.facets=TRUE)
}
