% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepared_examples.R
\name{prepared.examples}
\alias{prepared.examples}
\title{Prepared examples}
\usage{
prepared.examples(
  example.type = c("fcs.files.dt", "channel.alias", "dt", "fsom")
)
}
\arguments{
\item{example.type}{character string; matches one of: c('fcs.files.dt','channel.alias','dt','fsom')}
}
\value{
(hopefully) something useful for building examples and reducing redundancy
}
\description{
Prepared examples
}
\examples{
SOMnambulate:::prepared.examples(example.type='fcs.files.dt')[]
SOMnambulate:::prepared.examples(example.type='channel.alias')[]
SOMnambulate:::prepared.examples(example.type='dt')[]
str(SOMnambulate:::prepared.examples(example.type='fsom'))
}
