% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFCS_dt.R
\name{readFCS_dt}
\alias{readFCS_dt}
\title{Read a .fcs file and convert to a data.table}
\usage{
readFCS_dt(
  fcs.file.path,
  use.alias = T,
  use.alias.split = T,
  asinh.transform = F,
  cofactor.default = 1000,
  cofactor.mod = NULL,
  drop.events = T,
  natural.order = T,
  comp.mat.modified = NULL,
  channels.df = NULL,
  add.identifier.col = T
)
}
\arguments{
\item{fcs.file.path}{Character string; 'list.files(...,full.names=T,pattern=".fcs")'}

\item{use.alias}{Logical. By default, TRUE; will use header-encoded 'marker' names instead of 'channel' names.}

\item{use.alias.split}{Logical. By default, TRUE; will use 'split' alias names.}

\item{asinh.transform}{Logical. By default, FALSE; if TRUE, will transform flow/spectral/mass cytometry data using the function 'asinh()'}

\item{cofactor.default}{Adjust the 'asinh.transform' using a supplied cofactor.  Depending on cytometry data type, use the following default cofactor: flow = 1000, mass = 5}

\item{cofactor.mod}{Channel/marker-specific adjustment to 'cofactor.default'; provide as a named numeric vector, e.g. 'c("CD3"=2000,"CD8A"=2000)'}

\item{drop.events}{Logical. By default, TRUE; will drop 'extreme' events using the internal function 'drop_events_count_based()'; these events are most likely to be extreme-negative values introduced due to spectral compensation}

\item{natural.order}{Logical. By default, TRUE; will sort column names alphabetically then numerically, e.g. "CD3, CD4, CD8, CD45A" instead of "CD3, CD4, CD45A, CD8".}

\item{comp.mat.modified}{An optional compensation matrix; if provided, will be used in place of a header-encoded matrix ('$SPILL').}

\item{channels.df}{An optional data.frame returned from the function 'generate.channels.frame()'}

\item{add.identifier.col}{Logical. By default, TRUE; adds the .fcs filename (basename) as an identifier.}
}
\value{
A data.table composed of .fcs data (cells/events as rows; channels/markers as columns); appended with an additional identifier column ('sample.id').
}
\description{
Read a .fcs file and convert to a data.table
}
