% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.bivariate.shiny.R
\name{selection.heatmap}
\alias{selection.heatmap}
\title{Create a clickable \code{plotly::plot_ly} heatmap}
\usage{
selection.heatmap(vars, type = c("xy", "factor"))
}
\arguments{
\item{vars}{Character string ('variables') of class \code{numeric} or \code{factor}}

\item{type}{\code{xy} returns a clickable heatmap of numeric variables; \code{factor} returns a clickable heatmap of factored variables.}
}
\value{
a \code{plotly::plot_ly} heatmap; allows for clickable selection of reactive variables when used with \code{shiny}
}
\description{
Create a clickable \code{plotly::plot_ly} heatmap
}
