% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsom_funcs.R
\name{som.codes.dt}
\alias{som.codes.dt}
\title{Generate a \code{data.table} of clustered \code{FlowSOM} codes/SOMs.}
\usage{
som.codes.dt(
  fsom,
  append.name = NULL,
  k = NULL,
  umap.codes = F,
  seed.val = 1337
)
}
\arguments{
\item{fsom}{A \code{FlowSOM::SOM} result as returned from \link{som}.}

\item{append.name}{Character string; will be used to append the otherwise generically named 'node' and 'cluster' columns.}

\item{k}{Numeric; if defined, will generate \code{k} clusters using \link[FlowSOM]{metaClustering_consensus}.}

\item{umap.codes}{Logical: default \code{FALSE}; if \code{TRUE}, the codes/SOMs will be embedded using \link[uwot]{umap}.}

\item{seed.val}{An elite random seed value used to control otherwise random starts.}
}
\value{
A \code{data.table} of clustered (factor) codes that represents the direct result of \link[FlowSOM]{SOM}.
}
\description{
Generate a \code{data.table} of clustered \code{FlowSOM} codes/SOMs.
}
