% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMnambulate-package.R
\docType{package}
\name{SOMnambulate-package}
\alias{SOMnambulate}
\alias{SOMnambulate-package}
\title{SOMnambulate: Sleep Walking Through a High Dimensional Dream}
\description{
Generalized tools for .fcs preprocessing and visualization; High-dimensional analysis using a (flow)SOM-based workflow.
}
\author{
\strong{Maintainer}: Nathan Laniewski \email{nathan_laniewski@urmc.rochester.edu}

}
\keyword{internal}
