% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debarcode_funcs.R
\name{valleys.density.rug}
\alias{valleys.density.rug}
\title{Plot \code{fsom$codes} valleys}
\usage{
valleys.density.rug(fsom, out.name, plot.codes = NULL)
}
\arguments{
\item{fsom}{Object; the result of \code{FlowSOM::SOM(...)}}

\item{out.name}{Character string for use in plot title; usually the 'batch' name}

\item{plot.codes}{Numeric vector; if defined, will indicate expression level of specific codes}
}
\value{
\code{ggplot2} object
}
\description{
Plot \code{fsom$codes} valleys
}
