% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debarcode_funcs.R
\name{barcode.assignment.fsom.codes}
\alias{barcode.assignment.fsom.codes}
\title{Assign barcode values to the 'codes' result of a \code{FlowSOM::SOM} object based on an internally generated 'key'}
\usage{
barcode.assignment.fsom.codes(codes, m = 3)
}
\arguments{
\item{codes}{The 'codes' (SOMs) result of a \code{FlowSOM::SOM} object; \code{fsom$codes}}

\item{m}{Numeric. Default = 3. Argument passed to \code{utils::combn(x,m)}; used to define the barcode 'scheme': x-choose-m}
}
\value{
Numeric vector of barcode values with length equal to the number of SOMs
}
\description{
Based on the dimensions/markers used to barcode individual samples, this function will generate a unique barcode value for each SOM; the barcode value can then be assigned to the primary input data allowing for identification of sample-specific events/cells.
}
