% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cofactor_viewer.R
\name{cofactor.viewer}
\alias{cofactor.viewer}
\title{Interactive asinh transformation of cytometry data}
\usage{
cofactor.viewer(dat, cofactors.file.path = NULL)
}
\arguments{
\item{dat}{\code{data.table} containing raw, untransformed cytometry data}

\item{cofactors.file.path}{if provided, will save the cofactor values to the specified \code{file.path}}
}
\value{
\code{shinyapp}; if \code{cofactors.file.path} is provided, a \code{data.table} of cofactor values will be saved for use in transforming raw data values.
}
\description{
Interactive asinh transformation of cytometry data
}
