% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_funcs.R
\name{fcs.to.dt}
\alias{fcs.to.dt}
\title{reads a .fcs file and converts it to \code{data.table}}
\usage{
fcs.to.dt(
  fcs.dt,
  channel_alias = NULL,
  alias.pattern = F,
  alias.order = F,
  cofactors = NULL
)
}
\arguments{
\item{fcs.dt}{a \code{data.table} as returned from \code{get.fcs.file.dt}}

\item{channel_alias}{as returned from \code{get.fcs.channel.alias}}

\item{alias.pattern}{Logical; default \code{FALSE}. If \code{TRUE} and \code{channel_alias} is defined, the 'alias' column will be used as a pattern to define the \code{FlowCore::read.FCS(column.pattern = ...)} argument.}

\item{alias.order}{Logical. If \code{TRUE}, the \code{data.table} columns will be ordered to match that of the \code{channel_alias} 'alias' column.}

\item{cofactors}{A named numeric vector; names must match those found in \code{fcs.dt}; named columns will be \code{asinh} transformed with the supplied cofactor (numeric).}
}
\value{
a \code{data.table} of raw, un-transformed numeric expression values with character/factor identifier columns; if \code{cofactors} is defined, the raw expression values will be \code{asinh}-transformed.
}
\description{
reads a .fcs file and converts it to \code{data.table}
}
