% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_funcs.R
\name{get.fcs.channel.alias}
\alias{get.fcs.channel.alias}
\title{Get a \code{channel_alias} data.frame from .fcs file headers}
\usage{
get.fcs.channel.alias(fcs.file.paths)
}
\arguments{
\item{fcs.file.paths}{Character string; path(s) usually returned from \code{list.files(...,full.names=T,pattern=".fcs")}.}
}
\value{
returns a data.frame containing a 'channels' and 'alias' column
}
\description{
The resultant data.frame is to be used with the \code{channel_alias} argument of \code{flowCore::read.fcs(...)}
}
