% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_funcs.R
\name{get.fcs.file.dt}
\alias{get.fcs.file.dt}
\title{Get a \code{data.table} of .fcs file paths and related metadata.}
\usage{
get.fcs.file.dt(fcs.file.paths)
}
\arguments{
\item{fcs.file.paths}{Character string; path(s) usually returned from \code{list.files(...,full.names=T,pattern=".fcs")}.}
}
\value{
a \code{data.table} of full length .fcs file paths and related metadata (parsed from the text header).
}
\description{
Get a \code{data.table} of .fcs file paths and related metadata.
}
