% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_funcs.R
\name{get.fcs.keywords.metadata}
\alias{get.fcs.keywords.metadata}
\alias{getSOMkeywords}
\title{Get keyword metadata (!'$P') from .fcs files}
\usage{
get.fcs.keywords.metadata(fcs.file.paths, return.dt = T)

getSOMkeywords(fcs.file.paths, return.dt = T)
}
\arguments{
\item{fcs.file.paths}{Character string; path(s) usually returned from \code{list.files(...,full.names=T,pattern=".fcs")}.}

\item{return.dt}{Logical; default \code{TRUE}; if \code{FALSE}, will return a list of non-parameter (!'$P') keywords per .fcs file.}
}
\value{
A (row-bound) \link[data.table]{data.table} of keywords/values.
}
\description{
Non-parameter ($P) keywords in the TEXT section of a .fcs file denote cytometer/sample-specific meta-data. This function allows parsing of that meta-data without the overhead of reading the full (text and expression matrix) .fcs file to get the keyword values.  Depending on the nature of the .fcs file, there will be some mix of FCS-defined ('$') and non-standard keywords (software/user-defined).
}
\examples{
extdata<-system.file("extdata",package="SOMnambulate")
fcs.files <- list.files(extdata,full.names=TRUE,pattern=".fcs")

#a list of keywords/values (per .fcs file)
get.fcs.keywords.metadata(fcs.files[1],return.dt=FALSE)#a single file

#a (row-bound) data.table of keywords/values
mdat<- get.fcs.keywords.metadata(fcs.files)

#note column classes in the data.table; the TEXT section is encoded as character
mdat[]
str(mdat)

mdat[,class(`$DATE`)]#"character" class
mdat[,class(get('$DATE'))]#to avoid having to use back ticks

mdat[,'$DATE' := data.table::as.IDate(get('$DATE'),format="\%d-\%b-\%Y")]
mdat[,class(`$DATE`)]#"Date" class
mdat[,class(get('$DATE'))]#to avoid having to use back ticks

}
