% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsom_funcs.R
\name{map.som.data}
\alias{map.som.data}
\title{Map individual data points to nearest node.}
\usage{
map.som.data(fsom, dt)
}
\arguments{
\item{fsom}{An object as returned from \link{som}}

\item{dt}{An object as returned from \link{fcs.to.dt}}
}
\value{
This function modifies \code{dt} by reference using \code{data.table}'s \verb{:=} and \code{set} functions; two derived columns of 'mapped' data -- named generically as 'node' and 'node_dist' -- are added to \code{dt}. If \code{fsom} has been list-appended with the return of \link{som.codes.dt}, then an additional column -- 'cluster' -- will be added by reference.
}
\description{
This function is entirely dependent on the compiled C code found in (hidden function) \code{FlowSOM:::MapDataToCodes}; as an incredibly useful function, it is used here to map (often tens of millions) data points to their nearest node based on the trained SOM.  The result of \code{FlowSOM:::MapDataToCodes} is a two column matrix but in this function it is added to \code{dt} by 'reference' so that no copy is made.
}
\note{
Assignment operators \link[base]{<-} should not be used with this function as it modifies \code{dt} by reference; the intent is to avoid copying an often large \code{dt} object.
}
\references{
Van Gassen, Sofie, Britt Callebaut, Mary J Van Helden, Bart N Lambrecht, Piet Demeester, Tom Dhaene, and Yvan Saeys. 2015. “FlowSOM: Using Self-Organizing Maps for Visualization and Interpretation of Cytometry Data.” Cytometry Part A 87 (7): 636–45. https://onlinelibrary.wiley.com/doi/full/10.1002/cyto.a.22625.
}
