% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_funcs.R
\name{parms.adf.dt}
\alias{parms.adf.dt}
\title{Generate a \code{data.table} of .fcs parameters}
\usage{
parms.adf.dt(dt, type.convert = NULL, name.fix = NULL)
}
\arguments{
\item{dt}{\code{data.table} returned from \code{fcs.to.dt}; some mix of .fcs expression values and character/factor values.}

\item{type.convert}{Character vector; will attempt to convert each string element (named column in \code{dt}) to numeric}

\item{name.fix}{a \code{data.table} with two columns: 'name' and 'name.fix'; if 'name' matches, the value will be replaced with 'name.fix'.}
}
\value{
a \code{data.table} of .fcs parameters; used to define \code{flowCore::flowFrame(...,parameters = Biobase::AnnotatedDataFrame(...))}.
}
\description{
Generate a \code{data.table} of .fcs parameters
}
