% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsom_funcs.R
\name{som.codes.dt}
\alias{som.codes.dt}
\title{Generate a \link[data.table]{data.table} of clustered \link[FlowSOM:SOM]{FlowSOM::SOM} codes/SOMs.}
\usage{
som.codes.dt(
  fsom,
  append.name = NULL,
  k = NULL,
  umap.codes = F,
  seed.val = 1337,
  pItem = 1,
  reps = 100
)
}
\arguments{
\item{fsom}{A \link[FlowSOM:SOM]{FlowSOM::SOM} result as returned from \link{som}.}

\item{append.name}{Character string; will be used to append the otherwise generically named 'node' and 'cluster' columns.}

\item{k}{Numeric; if defined, will generate \code{k} clusters using \link[ConsensusClusterPlus]{ConsensusClusterPlus}.}

\item{umap.codes}{Logical: default \code{FALSE}; if \code{TRUE}, the codes/SOMs will be embedded using \link[uwot]{umap}.}

\item{seed.val}{An elite random seed value used to control otherwise random starts.}

\item{pItem}{Argument passed to \link[ConsensusClusterPlus]{ConsensusClusterPlus}; directly influences final clustering result.}

\item{reps}{Argument passed to \link[ConsensusClusterPlus]{ConsensusClusterPlus}; directly influences final clustering result.}
}
\value{
An updated \code{fsom}; the new list element \verb{[['codes.dt']]} will contain a \link[data.table]{data.table} of clustered (factor) codes using the direct result of \link[FlowSOM]{SOM}.
}
\description{
Using the direct result of \link{som}, the \verb{[['codes']]} are clustered and a column-bound \code{data.table} is generated.  As it is computationally expensive to UMAP-embed often millions of data points, an alternative is to embed the codes themselves for added interpretation of clustering results.
}
\examples{
fsom<-SOMnambulate:::prepared.examples(example.type='fsom')
fsom<-som.codes.dt(fsom,append.name = 'pbmc',k=10,umap.codes = TRUE)

head(fsom$codes.dt[])

ggplot2::ggplot(fsom$codes.dt,ggplot2::aes(umap.1,umap.2)) +
ggplot2::geom_point(ggplot2::aes(color=cluster_pbmc))

}
