% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMnambulate-package.R
\docType{package}
\name{SOMnambulate-package}
\alias{SOMnambulate}
\alias{SOMnambulate-package}
\title{SOMnambulate: Analysis and Visualization of High-Dimensional Mass/Spectral Cytometry Data Powered by FlowSOM Clustering}
\description{
SOMnambualte offers an efficient, reproducible, and scalable workflow for the analysis of high-dimensional mass/spectral cytometry data. It's like sleep walking through a high dimensional dream.
}
\author{
\strong{Maintainer}: Nathan Laniewski \email{nathan_laniewski@urmc.rochester.edu}

}
\keyword{internal}
